refactor(Dashboard): moved providers to separate file
